+++ breeze-unstable/src/main/kotlin/serializer/impl/BreezeSerializer.kt
@@ -1,0 +1,15 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.serializer.impl
+
+import com.windea.breezeframework.mapper.*
+import com.windea.breezeframework.serializer.*
+
+/**
+ * 由BreezeFramework实现的序列化器。
+ * @see com.windea.breezeframework.mapper.Mapper
+ * */
+interface BreezeSerializer<T : Mapper> : Serializer {
+	val delegate: T
+}