+++ breeze-unstable/src/main/kotlin/serializer/impl/JacksonSerializer.kt
@@ -1,0 +1,15 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.serializer.impl
+
+import com.fasterxml.jackson.databind.*
+import com.windea.breezeframework.serializer.*
+
+/**
+ * 由Jackson实现的序列化器。
+ * @see com.fasterxml.jackson.databind.ObjectMapper
+ * */
+interface JacksonSerializer<T : ObjectMapper> : Serializer {
+	val delegate: T
+}